Remove id from certain SVG elements.
Some of the SVG files have multiple elements with the same id.
This violates the specifications for the SVG format, and caused problems when I was trying to parse the files with another program.

Comparing these files to the other SVG sources for OpenMoji, I found that these id attributes are nonstandard.
Most of the SVG files have descriptive ids for each group ("line", "color", "skin", etc.), but not for the elements within those groups.
As such, the easiest fix was to simply remove those problematic id attributes.

These duplicate id attributes all have forms similar to `id="_x2935__xFE0F__1_"`.
While scanning the folders for ids of this form, I also found two av-symbol SVG files which had this form of id, but without the duplication.
Although these two files don't violate the specifications for SVG, I have also removed their ids for the sake of consistency.